feat(helper): AppShell — single header, files-first main region
Replace App.tsx's flat if-ladder of full-screen swaps with a persistent
AppShell that owns ONE header (drag region, traffic-light spacer, status
dot, title, a Files/Chat/History SegmentedControl, and the chat-scoped
Flag/New actions) over a single main region that swaps on mainView
(files | chat | history | device-info). App.tsx collapses to connection
gates → username gate → <AppShell/>; the showHistory/showDeviceInfo/
showWorkspace local state and the tray show-device-info listener move
into the shell. SessionHistory, DeviceInfoView, and the tool-approval
popup relocate from App.tsx; DeviceInfoView no longer emits a second
.helper-header.

WorkspacePanel gains an `embedded` prop (default false) that suppresses
its own header/Back and the Escape walk-back so the shell owns the header
and navigation; standalone rendering — and all existing W1 tests — stay
valid. Files nav and the files-first default appear only when the
workspace capability probe resolves available.
