+++ apps/helper/src/App.tsx
@@ -1,11 +1,7 @@
 import { useEffect, useState } from 'react';
-import { invoke } from '@tauri-apps/api/core';
-import { listen } from '@tauri-apps/api/event';
 import { useChatStore } from './stores/chatStore';
-import type { SessionSummary, PendingApproval, DeviceContext } from './stores/chatStore';
 import { useWorkspaceStore } from './stores/workspaceStore';
-import WorkspacePanel from './components/workspace/WorkspacePanel';
-import ChatView from './components/shell/ChatView';
+import AppShell from './components/shell/AppShell';
 
 function UsernamePrompt({ osUsername }: { osUsername?: string }) {
   const setUsername = useChatStore((s) => s.setUsername);
@@ -44,340 +40,17 @@ function UsernamePrompt({ osUsername }: { osUsername?: string }) {
   );
 }
 
-function formatDate(dateStr: string): string {
-  const d = new Date(dateStr);
-  const now = new Date();
-  const diffMs = now.getTime() - d.getTime();
-  const diffMins = Math.floor(diffMs / 60000);
-  if (diffMins < 1) return 'Just now';
-  if (diffMins < 60) return `${diffMins}m ago`;
-  const diffHours = Math.floor(diffMins / 60);
-  if (diffHours < 24) return `${diffHours}h ago`;
-  const diffDays = Math.floor(diffHours / 24);
-  if (diffDays < 7) return `${diffDays}d ago`;
-  return d.toLocaleDateString();
-}
-
-function SessionHistory({ onClose }: { onClose: () => void }) {
-  const { sessions, sessionsLoading, loadSession, loadSessions } = useChatStore();
-
-  useEffect(() => {
-    loadSessions();
-  }, [loadSessions]);
-
-  const handleSelect = (session: SessionSummary) => {
-    loadSession(session.id);
-    onClose();
-  };
-
-  return (
-    <div className="helper-history">
-      <div className="helper-history-header" data-tauri-drag-region>
-        {isMacOS && <div className="helper-traffic-light-spacer" />}
-        <span className="helper-history-title">History</span>
-        <div className="helper-header-drag-spacer" data-tauri-drag-region />
-        <button onClick={onClose} className="helper-btn helper-btn-sm">
-          Back
-        </button>
-      </div>
-      <div className="helper-history-list">
-        {sessionsLoading && (
-          <div className="helper-history-loading">
-            <span className="helper-spinner" />
-            <span>Loading...</span>
-          </div>
-        )}
-        {!sessionsLoading && sessions.length === 0 && (
-          <div className="helper-history-empty">No conversations yet</div>
-        )}
-        {sessions.map((s) => (
-          <button
-            key={s.id}
-            className="helper-history-item"
-            onClick={() => handleSelect(s)}
-          >
-            <span className="helper-history-item-title">
-              {s.title || 'Untitled'}
-            </span>
-            <span className="helper-history-item-meta">
-              {formatDate(s.updatedAt)}
-              {s.turnCount > 0 && ` · ${s.turnCount} turns`}
-            </span>
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-// Must be <= server-side waitForApproval timeout (300s). Plan approvals use 10-min timeout.
-const AUTO_DENY_MS = 5 * 60 * 1000; // 5 minutes
-const HIDDEN_INPUT_KEYS = new Set(['deviceId', 'orgId', 'siteId', 'sessionId']);
-
-function filterInput(input: Record<string, unknown>): Record<string, unknown> {
-  const filtered: Record<string, unknown> = {};
-  for (const [k, v] of Object.entries(input)) {
-    if (!HIDDEN_INPUT_KEYS.has(k)) filtered[k] = v;
-  }
-  return filtered;
-}
-
-function formatIdle(minutes: number): string {
-  if (minutes < 1) return 'just now';
-  if (minutes < 60) return `${minutes}m`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours}h`;
-  return `${Math.floor(hours / 24)}d`;
-}
-
-function formatDeviceIdle(lastSeenAt: string | undefined): string | null {
-  if (!lastSeenAt) return null;
-  const diffMs = Date.now() - new Date(lastSeenAt).getTime();
-  if (diffMs < 60_000) return null;
-  const mins = Math.floor(diffMs / 60_000);
-  if (mins < 60) return `${mins}m`;
-  const hours = Math.floor(mins / 60);
-  if (hours < 24) return `${hours}h`;
-  return `${Math.floor(hours / 24)}d`;
-}
-
-function DeviceBadge({ ctx }: { ctx: DeviceContext }) {
-  const name = ctx.displayName || ctx.hostname;
-  const isOnline = ctx.status === 'online';
-  const deviceIdleText = !isOnline ? formatDeviceIdle(ctx.lastSeenAt) : null;
-  const sessions = ctx.activeSessions ?? [];
-
-  return (
-    <div className="helper-approval-device">
-      <div className="helper-approval-device-row">
-        <span className="helper-approval-device-name">{name}</span>
-        <span className="helper-approval-device-sep">&middot;</span>
-        <span className={isOnline ? 'helper-approval-device-active' : 'helper-approval-device-idle'}>
-          {isOnline ? 'online' : (deviceIdleText ? `offline ${deviceIdleText}` : 'offline')}
-        </span>
-      </div>
-      {sessions.map((s, i) => {
-        const state = s.activityState ?? 'unknown';
-        const idleText = state !== 'active' && s.idleMinutes != null && s.idleMinutes > 0
-          ? `idle ${formatIdle(s.idleMinutes)}`
-          : state;
-        return (
-          <div key={i} className="helper-approval-device-session">
-            <span className="helper-approval-device-user">{s.username}</span>
-            {s.sessionType !== 'console' && (
-              <span className="helper-approval-device-type">{s.sessionType}</span>
-            )}
-            <span className={state === 'active' ? 'helper-approval-device-active' : 'helper-approval-device-idle'}>
-              {idleText}
-            </span>
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-function ToolApprovalPopup({
-  approval,
-  onApprove,
-  onDeny,
-}: {
-  approval: PendingApproval;
-  onApprove: () => void;
-  onDeny: () => void;
-}) {
-  const [remainingMs, setRemainingMs] = useState(AUTO_DENY_MS);
-
-  useEffect(() => {
-    const start = Date.now();
-    const interval = setInterval(() => {
-      const elapsed = Date.now() - start;
-      const remaining = AUTO_DENY_MS - elapsed;
-      if (remaining <= 0) {
-        clearInterval(interval);
-        onDeny();
-      } else {
-        setRemainingMs(remaining);
-      }
-    }, 1000);
-    return () => clearInterval(interval);
-  }, [onDeny]);
-
-  const minutes = Math.floor(remainingMs / 60000);
-  const seconds = Math.floor((remainingMs % 60000) / 1000);
-  const countdown = `${minutes}:${seconds.toString().padStart(2, '0')}`;
-
-  const visibleInput = filterInput(approval.input);
-  const hasVisibleInput = Object.keys(visibleInput).length > 0;
-
-  return (
-    <div className="helper-approval-overlay">
-      <div className="helper-approval-card">
-        <div className="helper-approval-header">
-          <span className="helper-approval-icon">&#9888;</span>
-          <span className="helper-approval-title">Approval Required</span>
-        </div>
-        <div className="helper-approval-body">
-          <div className="helper-approval-desc">{approval.description}</div>
-          {approval.deviceContext && <DeviceBadge ctx={approval.deviceContext} />}
-          {hasVisibleInput && (
-            <details className="helper-approval-details">
-              <summary>Show parameters</summary>
-              <pre className="helper-approval-input">
-                {JSON.stringify(visibleInput, null, 2)}
-              </pre>
-            </details>
-          )}
-        </div>
-        <div className="helper-approval-footer">
-          <span className="helper-approval-countdown">Auto-deny {countdown}</span>
-          <div className="helper-approval-actions">
-            <button onClick={onDeny} className="helper-btn helper-btn-deny">
-              Deny
-            </button>
-            <button onClick={onApprove} className="helper-btn helper-btn-allow">
-              Allow
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-type DeviceInfo = {
-  hostname: string;
-  osType: string;
-  osVersion: string;
-  status: string;
-  lastSeenAt?: string;
-  agentVersion?: string;
-};
-
-function DeviceInfoView({ onClose }: { onClose: () => void }) {
-  const [device, setDevice] = useState<DeviceInfo | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const { agentConfig } = useChatStore();
-  const helperVersion = agentConfig?.helper_version;
-
-  useEffect(() => {
-    if (!agentConfig) return;
-    setLoading(true);
-    setError(null);
-
-    invoke<{ status: number; body: string }>('helper_fetch', {
-        request: {
-          url: `${agentConfig.api_url}/api/v1/helper/device-info`,
-          method: 'GET',
-        },
-      })
-      .then((res) => {
-        if (res.status >= 200 && res.status < 300) {
-          const data = JSON.parse(res.body);
-          setDevice({
-            hostname: data.hostname || data.displayName || 'Unknown',
-            osType: data.osType || 'Unknown',
-            osVersion: data.osVersion || '',
-            status: data.status || 'unknown',
-            lastSeenAt: data.lastSeenAt,
-            agentVersion: data.agentVersion,
-          });
-        } else {
-          setError('Failed to load device info');
-        }
-      })
-      .catch((e: Error) => setError(e.message || 'Failed to load device info'))
-      .finally(() => setLoading(false));
-  }, [agentConfig]);
-
-  return (
-    <div className="helper-container">
-      <div className={`helper-header${isMacOS ? ' helper-header-macos' : ''}`} data-tauri-drag-region>
-        <div className="helper-header-left" data-tauri-drag-region>
-          {isMacOS && <div className="helper-traffic-light-spacer" />}
-          <span className="helper-title">Device Info</span>
-        </div>
-        <div className="helper-header-drag-spacer" data-tauri-drag-region />
-        <div className="helper-header-actions">
-          <button onClick={onClose} className="helper-btn helper-btn-sm">Back</button>
-        </div>
-      </div>
-      <div className="helper-messages" style={{ padding: '16px' }}>
-        {loading && (
-          <div className="helper-history-loading">
-            <span className="helper-spinner" />
-            <span>Loading...</span>
-          </div>
-        )}
-        {error && <div className="helper-error-banner"><span>{error}</span></div>}
-        {device && (
-          <div style={{ display: 'flex', flexDirection: 'column', gap: '12px' }}>
-            <div style={{ padding: '12px', border: '1px solid var(--helper-border, #e5e7eb)', borderRadius: '8px' }}>
-              <div style={{ fontSize: '13px', color: 'var(--helper-muted, #6b7280)' }}>Hostname</div>
-              <div style={{ fontSize: '15px', fontWeight: 600 }}>{device.hostname}</div>
-            </div>
-            <div style={{ padding: '12px', border: '1px solid var(--helper-border, #e5e7eb)', borderRadius: '8px' }}>
-              <div style={{ fontSize: '13px', color: 'var(--helper-muted, #6b7280)' }}>Operating System</div>
-              <div style={{ fontSize: '15px', fontWeight: 600 }}>{device.osType} {device.osVersion}</div>
-            </div>
-            <div style={{ padding: '12px', border: '1px solid var(--helper-border, #e5e7eb)', borderRadius: '8px' }}>
-              <div style={{ fontSize: '13px', color: 'var(--helper-muted, #6b7280)' }}>Status</div>
-              <div style={{ fontSize: '15px', fontWeight: 600, color: device.status === 'online' ? '#22c55e' : '#ef4444' }}>
-                {device.status}
-              </div>
-            </div>
-            {device.lastSeenAt && (
-              <div style={{ padding: '12px', border: '1px solid var(--helper-border, #e5e7eb)', borderRadius: '8px' }}>
-                <div style={{ fontSize: '13px', color: 'var(--helper-muted, #6b7280)' }}>Last Check-in</div>
-                <div style={{ fontSize: '15px', fontWeight: 600 }}>{formatDate(device.lastSeenAt)}</div>
-              </div>
-            )}
-            {device.agentVersion && (
-              <div style={{ padding: '12px', border: '1px solid var(--helper-border, #e5e7eb)', borderRadius: '8px' }}>
-                <div style={{ fontSize: '13px', color: 'var(--helper-muted, #6b7280)' }}>Agent Version</div>
-                <div style={{ fontSize: '15px', fontWeight: 600 }}>{device.agentVersion}</div>
-              </div>
-            )}
-            {helperVersion && (
-              <div style={{ padding: '12px', border: '1px solid var(--helper-border, #e5e7eb)', borderRadius: '8px' }}>
-                <div style={{ fontSize: '13px', color: 'var(--helper-muted, #6b7280)' }}>Helper Version</div>
-                <div style={{ fontSize: '15px', fontWeight: 600 }}>{helperVersion}</div>
-              </div>
-            )}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-const isMacOS = navigator.platform.startsWith('Mac') || navigator.userAgent.includes('Macintosh');
-
 export default function App() {
   const {
     connectionState,
     connectionError,
     agentConfig,
-    sessionId,
-    error,
     username,
-    pendingApproval,
-    isFlagged,
     initialize,
-    clearMessages,
-    approveExecution,
-    flagSession,
   } = useChatStore();
 
   const workspaceAvailable = useWorkspaceStore((s) => s.available);
   const probeWorkspace = useWorkspaceStore((s) => s.probe);
-
-  const [showHistory, setShowHistory] = useState(false);
-  const [showDeviceInfo, setShowDeviceInfo] = useState(false);
-  // Files-first: land in the Files view; the render below still gates on the
-  // capability probe, so orgs without workspace files keep the chat home.
-  const [showWorkspace, setShowWorkspace] = useState(true);
 
   useEffect(() => {
     initialize();
@@ -391,19 +64,7 @@ export default function App() {
     }
   }, [connectionState, workspaceAvailable, probeWorkspace]);
 
-  // Listen for tray menu "Device Info" click. Tray events only exist inside
-  // Tauri; the direct listen() import throws in browser dev mode, so skip it
-  // when the Tauri IPC bridge is absent.
-  useEffect(() => {
-    if (!('__TAURI_INTERNALS__' in window)) return;
-    let unlisten: (() => void) | undefined;
-    listen('show-device-info', () => {
-      setShowDeviceInfo(true);
-    }).then((fn: () => void) => { unlisten = fn; });
-    return () => { unlisten?.(); };
-  }, []);
-
-  // Connection states
+  // Connection states — full-screen gates above the shell.
   if (connectionState === 'connecting') {
     return (
       <div className="helper-container helper-center">
@@ -446,118 +107,12 @@ export default function App() {
     );
   }
 
-  // Username prompt (shown once before first use)
+  // Username prompt (shown once before first use) — full-screen gate.
   if (!username) {
     return <UsernamePrompt osUsername={agentConfig?.os_username} />;
   }
 
-  // Session history view
-  if (showHistory) {
-    return (
-      <div className="helper-container">
-        <SessionHistory onClose={() => setShowHistory(false)} />
-      </div>
-    );
-  }
-
-  // Device info view (triggered by tray menu)
-  if (showDeviceInfo) {
-    return <DeviceInfoView onClose={() => setShowDeviceInfo(false)} />;
-  }
-
-  // Workspace files view (only renders once the backend reports the capability)
-  if (showWorkspace && workspaceAvailable) {
-    return <WorkspacePanel onClose={() => setShowWorkspace(false)} />;
-  }
-
-  return (
-    <div className="helper-container bg-ws-canvas">
-      {/* Header — draggable title bar */}
-      <div className={`helper-header${isMacOS ? ' helper-header-macos' : ''}`} data-tauri-drag-region>
-        <div className="helper-header-left" data-tauri-drag-region>
-          {isMacOS && <div className="helper-traffic-light-spacer" />}
-          <span className="helper-status-dot helper-status-connected" />
-          <span className="helper-title">Breeze Helper</span>
-        </div>
-        <div className="helper-header-drag-spacer" data-tauri-drag-region />
-        <div className="helper-header-actions">
-          {sessionId && (
-            <button
-              onClick={() => flagSession('User flagged from helper')}
-              className={`helper-btn helper-btn-sm${isFlagged ? ' helper-btn-flagged' : ''}`}
-              title={isFlagged ? 'Conversation flagged' : 'Flag conversation for review'}
-              disabled={isFlagged}
-            >
-              {isFlagged ? 'Flagged' : 'Flag'}
-            </button>
-          )}
-          {workspaceAvailable === true && (
-            <button
-              onClick={() => setShowWorkspace(true)}
-              className="helper-btn helper-btn-sm"
-              title="Find your company's shared files"
-            >
-              Files
-            </button>
-          )}
-          <button
-            onClick={() => setShowHistory(true)}
-            className="helper-btn helper-btn-sm"
-            title="Conversation history"
-          >
-            History
-          </button>
-          <button
-            onClick={clearMessages}
-            className="helper-btn helper-btn-sm"
-            title="New conversation"
-          >
-            New
-          </button>
-          {!isMacOS && (
-            <>
-              <button
-                onClick={() => invoke('minimize_window').catch(() => {})}
-                className="helper-btn-window"
-                title="Minimize"
-              >
-                &#8211;
-              </button>
-              <button
-                onClick={() => invoke('hide_window').catch(() => {})}
-                className="helper-btn-window helper-btn-window-close"
-                title="Close to tray"
-              >
-                &#10005;
-              </button>
-            </>
-          )}
-        </div>
-      </div>
-
-      {/* Error banner */}
-      {error && (
-        <div className="helper-error-banner">
-          <span>{error}</span>
-          <button
-            onClick={() => useChatStore.setState({ error: null })}
-            className="helper-btn-close"
-          >
-            ×
-          </button>
-        </div>
-      )}
-
-      {/* Tool Approval Popup */}
-      {pendingApproval && (
-        <ToolApprovalPopup
-          approval={pendingApproval}
-          onApprove={() => approveExecution(pendingApproval.executionId, true)}
-          onDeny={() => approveExecution(pendingApproval.executionId, false)}
-        />
-      )}
-
-      <ChatView />
-    </div>
-  );
+  // Post-auth: the persistent files-first shell owns the header and all
+  // navigation (Files / Chat / History / Device Info).
+  return <AppShell />;
 }
